+++ dataland-e2etests/src/test/kotlin/org/dataland/e2etests/tests/MetaDataControllerTest.kt
@@ -10,7 +10,6 @@ import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
-import java.lang.IllegalArgumentException
 import java.math.BigDecimal
 import java.time.Instant
 import java.time.LocalDateTime
@@ -29,79 +28,15 @@ class MetaDataControllerTest {
     private val listOfTestCompanyInformation = apiAccessor.testDataProviderForEuTaxonomyDataForNonFinancials
         .getCompanyInformationWithoutIdentifiers(numberOfCompaniesToPostPerFramework)
     private val listOfOneTestCompanyInformation = listOf(listOfTestCompanyInformation[0])
-    private val listOfOneNonTeaserTestCompanyInformation =
-        listOf(listOfTestCompanyInformation[0].copy(isTeaserCompany = false))
-    private val listOfOneTeaserTestCompanyInformation =
-        listOf(listOfTestCompanyInformation[0].copy(isTeaserCompany = true))
 
-    private fun validateAdminAccessToUserId(
-        testUploadDataUploaderMetaInfo: DataMetaInformation,
-        testUploadDataAdminMetaInfo: DataMetaInformation,
-    ) {
-        expectUserIdToBe(
-            testUploadDataUploaderMetaInfo, TechnicalUser.Admin, TechnicalUser.Uploader.technicalUserId,
-            "Admins should be able to view uploaderUserids for all users",
-        )
-        expectUserIdToBe(
-            testUploadDataAdminMetaInfo, TechnicalUser.Admin, TechnicalUser.Admin.technicalUserId,
-            "Admins should be able to view uploaderUserids for all users",
-        )
-    }
-
-    private fun validateUploaderAccessToUserId(
-        testUploadDataUploaderMetaInfo: DataMetaInformation,
-        testUploadDataAdminMetaInfo: DataMetaInformation,
-    ) {
-        expectUserIdToBe(
-            testUploadDataUploaderMetaInfo, TechnicalUser.Uploader, TechnicalUser.Uploader.technicalUserId,
-            "Expected user id to be present if the user requests data about an upload he performed himself",
-        )
-        expectUserIdToBe(
-            testUploadDataAdminMetaInfo, TechnicalUser.Uploader, null,
-            "Data Uploaders should not be able to view the user id of uploads of other users",
-        )
-    }
-
-    private fun validateReaderAccessToUserId(
-        testUploadDataUploaderMetaInfo: DataMetaInformation,
-        testUploadDataAdminMetaInfo: DataMetaInformation,
-    ) {
-        expectUserIdToBe(
-            testUploadDataUploaderMetaInfo, TechnicalUser.Reader, null,
-            "A reader should not see any uploader ids",
-        )
-        expectUserIdToBe(
-            testUploadDataAdminMetaInfo, TechnicalUser.Reader, null,
-            "A reader should not see any uploader ids",
-        )
-    }
-
-    private fun expectUserIdToBe(
-        dataMetaInformation: DataMetaInformation,
-        requestingTechnicalUser: TechnicalUser,
-        expectedUploaderId: String?,
-        msg: String,
-    ) {
-        apiAccessor.jwtHelper.authenticateApiCallsWithJwtForTechnicalUser(requestingTechnicalUser)
-
-        val uploaderUserIdFromMetaInfo = apiAccessor.metaDataControllerApi.getDataMetaInfo(dataMetaInformation.dataId)
-            .uploaderUserId
-        assertEquals(expectedUploaderId, uploaderUserIdFromMetaInfo, msg)
-
-        val uploaderUserIdFromCompanyInfo = apiAccessor.companyDataControllerApi
-            .getCompanyById(dataMetaInformation.companyId)
-            .dataRegisteredByDataland.firstOrNull()?.uploaderUserId
-        assertEquals(uploaderUserIdFromCompanyInfo, uploaderUserIdFromMetaInfo, msg)
-    }
-
-    private fun buildAcceptedAndActiveDataMetaInformation(
+    fun buildAcceptedAndActiveDataMetaInformation(
         dataId: String,
         companyId: String,
         testDataType: DataTypeEnum,
         uploadTime: Long,
     ) = DataMetaInformation(
-        dataId, companyId, testDataType, uploadTime,
-        "", true, QAStatus.accepted, null,
+        dataId = dataId, companyId = companyId, dataType = testDataType, uploadTime = uploadTime,
+        reportingPeriod = "", currentlyActive = true, qaStatus = QAStatus.accepted, uploaderUserId = null,
     )
 
     @Test
@@ -147,7 +82,7 @@ class MetaDataControllerTest {
         assertEquals(
             expectedSizeOfDataMetaInfo, sizeOfListOfDataMetaInfo,
             "The list with all data meta info is expected to increase by $totalNumberOfDataSetsPerFramework to " +
-                    "$expectedSizeOfDataMetaInfo, but has the size $sizeOfListOfDataMetaInfo.",
+                "$expectedSizeOfDataMetaInfo, but has the size $sizeOfListOfDataMetaInfo.",
         )
     }
 
@@ -165,7 +100,7 @@ class MetaDataControllerTest {
         assertEquals(
             numberOfDataSetsToPostPerCompany, listOfDataMetaInfoForFirstCompanyId.size,
             "The first posted company is expected to have meta info about $numberOfDataSetsToPostPerCompany " +
-                    "data sets, but has meta info about ${listOfDataMetaInfoForFirstCompanyId.size} data sets.",
+                "data sets, but has meta info about ${listOfDataMetaInfoForFirstCompanyId.size} data sets.",
         )
     }
 
@@ -186,7 +121,7 @@ class MetaDataControllerTest {
         assertEquals(
             expectedListSize, listSizeAfterUploads,
             "The meta info list for all EU Taxonomy Data for Non-Financials is expected to increase by " +
-                    "$totalNumberOfDataSetsPerFramework to $expectedListSize, but has the size $listSizeAfterUploads.",
+                "$totalNumberOfDataSetsPerFramework to $expectedListSize, but has the size $listSizeAfterUploads.",
         )
     }
 
@@ -204,92 +139,7 @@ class MetaDataControllerTest {
         assertEquals(
             numberOfDataSetsToPostPerCompany, sizeOfListOfDataMetaInfoPerCompanyIdAndDataType,
             "The first posted company is expected to have meta info about $numberOfDataSetsToPostPerCompany " +
-                    "data sets, but has meta info about $sizeOfListOfDataMetaInfoPerCompanyIdAndDataType data sets.",
-        )
-    }
-
-    @Test
-    fun `post a dummy teaser company and data for it and confirm unauthorized meta info access succeeds`() {
-        val testDataType = DataTypeEnum.eutaxonomyMinusFinancials
-        val listOfUploadInfo = apiAccessor.uploadCompanyAndFrameworkDataForMultipleFrameworks(
-            mapOf(testDataType to listOfOneTeaserTestCompanyInformation), 1,
-        )
-        val testDataId = listOfUploadInfo[0].actualStoredDataMetaInfo!!.dataId
-        val dataMetaInformation = apiAccessor.unauthorizedMetaDataControllerApi.getDataMetaInfo(testDataId)
-        val uploadTime = LocalDateTime.now().toEpochSecond(ZoneOffset.UTC)
-        assertEquals(
-            buildAcceptedAndActiveDataMetaInformation(
-                testDataId,
-                listOfUploadInfo[0].actualStoredCompany.companyId,
-                testDataType,
-                uploadTime,
-            ),
-            dataMetaInformation.copy(uploadTime = uploadTime),
-            "The meta info of the posted eu taxonomy data does not match the retrieved meta info.",
-        )
-    }
-
-    @Test
-    fun `post a dummy company and taxonomy data for it and confirm unauthorized meta info access is denied`() {
-        val listOfUploadInfo = apiAccessor.uploadCompanyAndFrameworkDataForMultipleFrameworks(
-            mapOf(DataTypeEnum.eutaxonomyMinusFinancials to listOfOneNonTeaserTestCompanyInformation), 1,
-        )
-        val testDataId = listOfUploadInfo[0].actualStoredDataMetaInfo!!.dataId
-        val exception = assertThrows<IllegalArgumentException> {
-            apiAccessor.unauthorizedMetaDataControllerApi.getDataMetaInfo(testDataId)
-        }
-        assertTrue(exception.message!!.contains("Unauthorized access failed"))
-    }
-
-    @Test
-    fun `post a dummy company as teaser company and data for it and confirm unauthorized meta info search succeeds`() {
-        val testDataType = DataTypeEnum.eutaxonomyMinusFinancials
-        val listOfUploadInfo = apiAccessor.uploadCompanyAndFrameworkDataForMultipleFrameworks(
-            mapOf(testDataType to listOfOneTeaserTestCompanyInformation), 1,
-        )
-        val testDataId = listOfUploadInfo[0].actualStoredDataMetaInfo!!.dataId
-        val testCompanyId = listOfUploadInfo[0].actualStoredCompany.companyId
-        val uploadTime = LocalDateTime.now().toEpochSecond(ZoneOffset.UTC)
-        val expectedMetaInformation = buildAcceptedAndActiveDataMetaInformation(
-            dataId = testDataId, companyId = testCompanyId,
-            testDataType = testDataType,
-            uploadTime = uploadTime,
-        )
-        assertTrue(
-            apiAccessor.unauthorizedMetaDataControllerApi.getListOfDataMetaInfo(testCompanyId, testDataType)
-                .map { it.copy(uploadTime = uploadTime) }.contains(expectedMetaInformation),
-            "The meta info of the posted eu taxonomy data that was associated with the teaser company " +
-                    "does not match the retrieved meta info.",
-        )
-    }
-
-    @Test
-    fun `post a dummy company and taxonomy data for it and confirm unauthorized meta info search is denied`() {
-        val testDataType = DataTypeEnum.eutaxonomyMinusFinancials
-        val testCompanyId = apiAccessor.uploadCompanyAndFrameworkDataForMultipleFrameworks(
-            mapOf(testDataType to listOfOneNonTeaserTestCompanyInformation), 1,
-        )[0].actualStoredCompany.companyId
-        val exception = assertThrows<IllegalArgumentException> {
-            apiAccessor.unauthorizedMetaDataControllerApi.getListOfDataMetaInfo(testCompanyId, testDataType)
-        }
-        assertTrue(exception.message!!.contains("Unauthorized access failed"))
-    }
-
-    @Test
-    fun `post two companies with data and check that the access to the uploaderUserId field is restricted`() {
-        val testDataType = DataTypeEnum.eutaxonomyMinusFinancials
-        val metaInfoOfUploaderUpload = apiAccessor.uploadCompanyAndFrameworkDataForMultipleFrameworks(
-            mapOf(testDataType to listOfOneNonTeaserTestCompanyInformation), 1, TechnicalUser.Uploader,
-        )[0].actualStoredDataMetaInfo!!
-        val metaInfoOfAdminUpload = apiAccessor.uploadCompanyAndFrameworkDataForMultipleFrameworks(
-            mapOf(testDataType to listOfOneNonTeaserTestCompanyInformation), 1, TechnicalUser.Admin,
-        )[0].actualStoredDataMetaInfo!!
-
-        validateReaderAccessToUserId(metaInfoOfUploaderUpload, metaInfoOfAdminUpload)
-        validateUploaderAccessToUserId(metaInfoOfUploaderUpload, metaInfoOfAdminUpload)
-        validateAdminAccessToUserId(
-            metaInfoOfUploaderUpload,
-            metaInfoOfAdminUpload,
+                "data sets, but has meta info about $sizeOfListOfDataMetaInfoPerCompanyIdAndDataType data sets.",
         )
     }
 
@@ -302,13 +152,17 @@ class MetaDataControllerTest {
     private fun ensureThatSecondDatasetIsActive(
         companyId: String,
         reportingPeriod: String,
-        newNumberOfEmployees: BigDecimal
+        newNumberOfEmployees: BigDecimal,
     ) {
         val dataType = DataTypeEnum.eutaxonomyMinusNonMinusFinancials
         val allDatasets =
-            apiAccessor.metaDataControllerApi.getListOfDataMetaInfo(companyId, dataType, false, reportingPeriod)
+            apiAccessor.metaDataControllerApi.getListOfDataMetaInfo(
+                companyId = companyId, dataType = dataType, showOnlyActive = false, reportingPeriod = reportingPeriod,
+            )
         val activeDatasets =
-            apiAccessor.metaDataControllerApi.getListOfDataMetaInfo(companyId, dataType, true, reportingPeriod)
+            apiAccessor.metaDataControllerApi.getListOfDataMetaInfo(
+                companyId = companyId, dataType = dataType, showOnlyActive = true, reportingPeriod = reportingPeriod,
+            )
         assertEquals(
             2, allDatasets.size, "The number of versions does not equal the expected one.",
         )
@@ -341,11 +195,11 @@ class MetaDataControllerTest {
         val newNumberOfEmployees = (frameworkDataAlpha.numberOfEmployees ?: BigDecimal.ZERO) + BigDecimal.ONE
         val frameworkDataBeta = frameworkDataAlpha.copy(numberOfEmployees = newNumberOfEmployees)
         apiAccessor.uploadSingleFrameworkDataSet(
-                companyId = companyId,
-                frameworkData = frameworkDataBeta,
-                reportingPeriod = reportingPeriod,
-                frameworkDataUploadFunction = apiAccessor.euTaxonomyNonFinancialsUploaderFunction,
-            )
+            companyId = companyId,
+            frameworkData = frameworkDataBeta,
+            reportingPeriod = reportingPeriod,
+            frameworkDataUploadFunction = apiAccessor.euTaxonomyNonFinancialsUploaderFunction,
+        )
         return Triple(companyId, reportingPeriod, newNumberOfEmployees)
     }
 
@@ -389,8 +243,10 @@ class MetaDataControllerTest {
         )
         uploadPairs.forEach { pair ->
             apiAccessor.uploadWithWait(
-                companyId, pair.first, pair.second,
-                apiAccessor.euTaxonomyNonFinancialsUploaderFunction,
+                companyId = companyId,
+                frameworkData = pair.first,
+                reportingPeriod = pair.second,
+                uploadFunction = apiAccessor.euTaxonomyNonFinancialsUploaderFunction,
             )
         }
         return companyId
